Remove support for hmac-sha1
The previous version of hmac-sha1 had a key confusion vulnerability,
where an attacker could trick someone expected a document to
be signed using a public key method, to accept a document that
has an HMAC signature with the public key as its HMAC secret.

Since SHA-1 is discouraged anyways, just remove support for HMAC-SHA1.

https://app.asana.com/0/261095544/1192755934099855/f